backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
neo4j>=5.15.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
